refactor(channels): hoist reply protocol out of discord, move silent-reply into channels/
- channels/discord/reply-directive.ts → channels/reply.ts (channel-agnostic)
- silent-reply.ts → channels/no-reply.ts
- update imports in discord/inbound, discord/outbound, automation/heartbeat
